fixed bugs

+++ include/ExecutorManager.h
@@ -19,7 +19,7 @@ class ExecutorManager :
 public:
 
     Executor*               find( ClusterSession * session );
-    Executor*               find( int session_id );
+    Executor*               find( size_t session_id );
     Executor*               find( std::string executor_id );
     std::vector<Executor*>  find_by_taskid( std::string task_id );
